requirements.txt
httpx==0.22.0
pytest==7.3.1
pytest-asyncio==0.17.0
python-dotenv==0.19.2
pytest-html==3.2.0